add new discretisation schemes
